Created shell script to post files in dir
The shell script will look for files in a directory and post ones it
hasn't posted before. It uses a hidden sqlite database in the directory
to determine which files it has already posted. Files are tracked by
their content so renaming a file will not caause it to be uploaded
again. However, the script will not currently spot cases where a file
is renamed and if it hasn't uploaded the file yet it will not use the
new name.