Made scripts more robust and backward compatible
The Exception as ex syntax doesn't work in an older version of Python,
so it was changed to the older Exception, ex syntax.

The post.sh was getting errors when using the while read; do; <<<
syntax as it would get an empty line on no input so I updated it to
check for empty lines and continue.

A GOOGLE_SITE environment variable can now be passed to post.sh which
will make it pass it as an argument to the Python scripts.